Captcha generation & user ids changed to bigint

+++ migrations/2024-12-13-151546_users/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    id            INTEGER UNSIGNED NOT NULL AUTO_INCREMENT,
+    id            BIGINT  UNSIGNED NOT NULL AUTO_INCREMENT,
     access_level  TINYINT UNSIGNED NOT NULL,
     username      VARCHAR(16)      UNIQUE,
     email         VARCHAR(128)     UNIQUE,
